\name{calc.ta}
\alias{calc.ta}
\title{ Calculate achievement function for the k-th priority level }
\description{
  This function calculcates the achievement function for the k-th priority level.
}
\usage{
calc.ta(tab, k)
}
\arguments{
  \item{tab}{ An object of class 'llgptab' that is the modified simplex tableau }
  \item{k}{ an integer priority level }
}
\value{
  None.
}
\references{
Ignizio, J. P. (1976). Goal Programming and Extensions, Lexington Books, D. C. Heath and Company.
}
\author{ Frederick Novomestky \email{fnovomes@poly.edu} }
\seealso{
\code{\link{llgptab}}
}
\keyword{ math }
