\name{calc.ti}
\alias{calc.ti}
\title{ Calculate the k-th index row }
\description{
  This function calculates the index row for the k-th priority level.
}
\usage{
calc.ti(tab, k)
}
\arguments{
  \item{tab}{ An object of class 'llgptab' that is the modified simplex tableau }
  \item{k}{ An integer priority level }
}
\value{
  None.
}
\references{
Ignizio, J. P. (1976). Goal Programming and Extensions, Lexington Books, D. C. Heath and Company.
}
\author{ Frederick Novomestky \email{fnovomes@poly.edu} }
\seealso{
\code{\link{llgptab}}
}
\keyword{ math }
