\name{coefficients}
\docType{data}
\alias{coefficients}
\title{Ignizio (1976) Example Data Sets}
\description{
  The data set is a matrix of coefficients for
  the linear objective functions for a linear goal programming
  problem.  The number of rows equals the number of
  objectives and the number columns equals the number
  of decision variables.
}
\format{
  The data set is a matrix.
} 
\references{
  Ignizio, J. P. (1976). Goal Programming and Extensions, Lexington Books.
} 
\author{ Frederick Novomestky \email{fnovomes@poly.edu} }
\seealso{
\code{\link{ignizio.datasets}}
}
\keyword{datasets}
