\name{ignizio.datasets}
\docType{data}
\alias{ignizio.datasets}
\title{Ignizio (1976) Sample Data Sets}
\description{
  A collection and description of data sets that
  are examples found in Ignizio (1976).  These are
  examples of goal programming problems solved
  using the methods described by Ignizio.

}
\format{
  Each data set is an R file that creates the coefficients matrix,
  the vector of target values and the data frame of achievement
  goals.
} 
\references{
  Ignizio, J. P. (1976). Goal Programming and Extensions, Lexington Books.
}
\author{ Frederick Novomestky \email{fnovomes@poly.edu} }
\seealso{
\code{\link{ignizio.example.3.1}},
\code{\link{ignizio.example.3.2}},
\code{\link{ignizio.example.3.3}},
\code{\link{ignizio.example.3.5}},
\code{\link{ignizio.example.3.6}},
\code{\link{llgp}}
}
\keyword{datasets}
