\name{targets}
\docType{data}
\alias{targets}
\title{Ignizio (1976) Example Data Sets}
\description{
  The data set is a vector of target values for the objectives.
  The number of elements is equal to the number of objectives.
}
\format{
  The data set is a vector.
} 
\references{
  Ignizio, J. P. (1976). Goal Programming and Extensions, Lexington Books.
} 
\author{ Frederick Novomestky \email{fnovomes@poly.edu} }
\seealso{
\code{\link{ignizio.datasets}}
}
\keyword{datasets}
